project-RAG/requirements.txt
streamlit
sentence-transformers
groq
requests
numpy
python-dotenv
msgpack
